+++ src/commands/db_cmds/db.rs
@@ -1,4 +1,7 @@
-use super::persistent::{path_to_db, path_to_env};
+use super::{
+    db_errors,
+    persistent::{path_to_db, path_to_env},
+};
 use chrono::Utc;
 use rusqlite::{params, Connection, Result};
 
@@ -17,11 +20,18 @@ pub struct DB {
 }
 
 impl DB {
-    pub fn new() -> Result<Self> {
+    pub fn new() -> Result<Self, Box<dyn std::error::Error>> {
         let env_path = path_to_env();
         dotenv::from_path(env_path).expect("Failed to read .env file");
         let curr_db = dotenv::var("db_var").expect("Unable to find db_var in .env file");
-        let path_to_db = path_to_db(&curr_db);
+        if curr_db == "none" {
+            return Err(db_errors::DatabaseError::new(
+                "No database selected",
+                db_errors::DatabaseErrorKind::VariableNotSet,
+            )
+            .into());
+        }
+        let path_to_db = path_to_db(&curr_db).expect("Failed to find db file");
         let conn = Connection::open(path_to_db)?; // Open or create a SQLite database file
         conn.execute(
             "CREATE TABLE IF NOT EXISTS tasks (